Fix config tests
They were not properly separated previously, and were reading from
each other's configs. This is because Pytest uses unique directory
names based on the function name, but it truncates after a certain
number of letters. If the first N letters are all the same, they will
all use the same directory. On top of this, I was not setting the
environment properly. That should now all be fixed.